Removed KaldbRpcUtil.
