--- kaldb/src/test/java/com/slack/kaldb/testlib/KaldbRpcUtil.java
@@ -1,78 +1,0 @@
-package com.slack.kaldb.testlib;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.linecorp.armeria.client.Clients;
-import com.slack.kaldb.proto.config.KaldbConfigs;
-import com.slack.kaldb.proto.service.KaldbSearch;
-import com.slack.kaldb.proto.service.KaldbServiceGrpc;
-import java.io.IOException;
-import java.util.HashMap;
-import org.apache.http.HttpEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class KaldbRpcUtil {
-  private static final Logger LOG = LoggerFactory.getLogger(KaldbRpcUtil.class);
-
-  public static KaldbSearch.SearchResult searchUsingGrpcApi(
-      String queryString,
-      long startTimeMs,
-      long endTimeMs,
-      String indexName,
-      int howMany,
-      int bucketCount,
-      int port) {
-    KaldbServiceGrpc.KaldbServiceBlockingStub kaldbService =
-        Clients.newClient(uri(port), KaldbServiceGrpc.KaldbServiceBlockingStub.class);
-
-    return kaldbService.search(
-        KaldbSearch.SearchRequest.newBuilder()
-            .setIndexName(indexName)
-            .setQueryString(queryString)
-            .setStartTimeEpochMs(startTimeMs)
-            .setEndTimeEpochMs(endTimeMs)
-            .setHowMany(howMany)
-            .setBucketCount(bucketCount)
-            .build());
-  }
-
-  private static String uri(int port) {
-    return "gproto+http://127.0.0.1:" + port + '/';
-  }
-
-  private static String getHealthCheckResponse(String url) {
-    try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-      HttpGet httpGet = new HttpGet(url);
-      try (CloseableHttpResponse httpResponse = httpclient.execute(httpGet)) {
-        HttpEntity entity = httpResponse.getEntity();
-
-        String response = EntityUtils.toString(entity);
-        EntityUtils.consume(entity);
-        return response;
-      }
-    } catch (IOException e) {
-      return null;
-    }
-  }
-
-  private static String getHealthCheckResponse(int port) {
-    String url = String.format("http://localhost:%s/health", port);
-    return getHealthCheckResponse(url);
-  }
-
-  public static boolean runHealthCheckOnPort(KaldbConfigs.ServerConfig serverConfig)
-      throws JsonProcessingException {
-    final ObjectMapper om = new ObjectMapper();
-    final String response = getHealthCheckResponse(serverConfig.getServerPort());
-    HashMap<String, Object> map = om.readValue(response, HashMap.class);
-
-    LOG.info(String.format("Response from healthcheck - '%s'", response));
-    return (boolean) map.get("healthy");
-  }
-}
